+++ src/platform_impl/linux/wayland/seat/pointer/relative_pointer.rs
@@ -72,6 +72,20 @@ impl Dispatch<ZwpRelativePointerV1, GlobalData, WinitState> for RelativePointerS
                 },
                 super::DeviceId,
             );
+            state.events_sink.push_device_event(
+                DeviceEvent::Motion {
+                    axis: 0,
+                    value: dx_unaccel,
+                },
+                super::DeviceId,
+            );
+            state.events_sink.push_device_event(
+                DeviceEvent::Motion {
+                    axis: 1,
+                    value: dy_unaccel,
+                },
+                super::DeviceId,
+            );
         }
     }
 }